remove unused tests
