--- sdk/src/test/java/com/hedera/hashgraph/sdk/HookUtilitiesTest.java
@@ -1,328 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-package com.hedera.hashgraph.sdk;
-
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
-
-import java.util.Arrays;
-import java.util.List;
-
-/**
- * Comprehensive test suite for hook utility classes.
- */
-public class HookUtilitiesTest {
-
-    @Test
-    public void testHookValidator() {
-        // Test valid hook creation details
-        ContractId contractId = new ContractId(100);
-        EvmHookSpec evmHookSpec = new EvmHookSpec(contractId);
-        LambdaEvmHook lambdaEvmHook = new LambdaEvmHook(evmHookSpec);
-        HookCreationDetails hookDetails = new HookCreationDetails(
-                HookExtensionPoint.ACCOUNT_ALLOWANCE_HOOK, 1L, lambdaEvmHook);
-
-        // Should not throw
-        HookValidator.validateHookCreationDetails(List.of(hookDetails), HookValidator.HookContext.ACCOUNT_CREATION);
-
-        // Test duplicate hook IDs
-        HookCreationDetails duplicateHook = new HookCreationDetails(
-                HookExtensionPoint.ACCOUNT_ALLOWANCE_HOOK, 1L, lambdaEvmHook);
-
-        assertThrows(IllegalArgumentException.class, () -> {
-            HookValidator.validateHookCreationDetails(List.of(hookDetails, duplicateHook), HookValidator.HookContext.ACCOUNT_CREATION);
-        });
-
-        // Test null hook details
-        assertThrows(IllegalArgumentException.class, () -> {
-            HookValidator.validateHookCreationDetails(null, HookValidator.HookContext.ACCOUNT_CREATION);
-        });
-
-        // Test null elements
-        assertThrows(IllegalArgumentException.class, () -> {
-            HookValidator.validateHookCreationDetails(Arrays.asList(hookDetails, null), HookValidator.HookContext.ACCOUNT_CREATION);
-        });
-    }
-
-    @Test
-    public void testHookValidatorStorageUpdates() {
-        // Test valid storage updates
-        LambdaStorageUpdate storageSlot = new LambdaStorageUpdate.LambdaStorageSlot(new byte[]{0x01}, new byte[]{0x02});
-        LambdaMappingEntry entry = LambdaMappingEntry.ofKey(new byte[]{0x03}, new byte[]{0x04});
-        LambdaStorageUpdate mappingEntries = new LambdaStorageUpdate.LambdaMappingEntries(new byte[]{0x05}, List.of(entry));
-
-        List<LambdaStorageUpdate> updates = List.of(storageSlot, mappingEntries);
-        HookValidator.validateStorageUpdates(updates);
-
-        // Test empty storage updates
-        assertThrows(IllegalArgumentException.class, () -> {
-            HookValidator.validateStorageUpdates(List.of());
-        });
-
-        // Test null storage updates
-        assertThrows(IllegalArgumentException.class, () -> {
-            HookValidator.validateStorageUpdates(null);
-        });
-
-        // Test too many storage updates
-        List<LambdaStorageUpdate> tooManyUpdates = new java.util.ArrayList<>();
-        for (int i = 0; i < 1001; i++) {
-            tooManyUpdates.add(new LambdaStorageUpdate.LambdaStorageSlot(new byte[]{0x01}, new byte[]{0x02}));
-        }
-        assertThrows(IllegalArgumentException.class, () -> {
-            HookValidator.validateStorageUpdates(tooManyUpdates);
-        });
-    }
-
-    @Test
-    public void testHookBuilder() {
-        ContractId contractId = new ContractId(200);
-        Key adminKey = PrivateKey.generateED25519().getPublicKey();
-
-        // Test basic hook building
-        HookCreationDetails hookDetails = HookBuilder.accountAllowanceHook()
-                .setHookId(1L)
-                .setContract(contractId)
-                .setAdminKey(adminKey)
-                .addStorageSlot(new byte[]{0x01}, new byte[]{0x02})
-                .build();
-
-        assertEquals(HookExtensionPoint.ACCOUNT_ALLOWANCE_HOOK, hookDetails.getExtensionPoint());
-        assertEquals(1L, hookDetails.getHookId());
-        assertEquals(contractId, hookDetails.getHook().getSpec().getContractId());
-        assertEquals(adminKey, hookDetails.getAdminKey());
-        assertEquals(1, hookDetails.getHook().getStorageUpdates().size());
-
-        // Test validation
-        HookCreationDetails validatedHook = HookBuilder.accountAllowanceHook()
-                .setHookId(2L)
-                .setContract(contractId)
-                .buildAndValidate(HookValidator.HookContext.ACCOUNT_CREATION);
-
-        assertNotNull(validatedHook);
-
-        // Test missing required fields
-        assertThrows(IllegalStateException.class, () -> {
-            HookBuilder.accountAllowanceHook().build();
-        });
-
-        // Test mapping entries
-        HookCreationDetails mappingHook = HookBuilder.accountAllowanceHook()
-                .setHookId(3L)
-                .setContract(contractId)
-                .addMappingEntry(new byte[]{0x10}, new byte[]{0x11}, new byte[]{0x12})
-                .addMappingEntryWithPreimage(new byte[]{0x13}, new byte[]{0x14, 0x15}, new byte[]{0x16})
-                .build();
-
-        assertEquals(2, mappingHook.getHook().getStorageUpdates().size());
-    }
-
-    @Test
-    public void testHookUtils() {
-        ContractId contractId = new ContractId(300);
-
-        // Test simple hook creation
-        HookCreationDetails simpleHook = HookUtils.createSimpleAccountAllowanceHook(1L, contractId);
-        assertEquals(HookExtensionPoint.ACCOUNT_ALLOWANCE_HOOK, simpleHook.getExtensionPoint());
-        assertEquals(1L, simpleHook.getHookId());
-        assertEquals(contractId, simpleHook.getHook().getSpec().getContractId());
-        assertNull(simpleHook.getAdminKey());
-
-        // Test hook with admin key
-        Key adminKey = PrivateKey.generateED25519().getPublicKey();
-        HookCreationDetails adminHook = HookUtils.createAccountAllowanceHookWithAdmin(2L, contractId, adminKey);
-        assertEquals(adminKey, adminHook.getAdminKey());
-
-        // Test hook with storage
-        LambdaStorageUpdate storageUpdate = new LambdaStorageUpdate.LambdaStorageSlot(new byte[]{0x01}, new byte[]{0x02});
-        HookCreationDetails storageHook = HookUtils.createHookWithStorage(3L, contractId, List.of(storageUpdate));
-        assertEquals(1, storageHook.getHook().getStorageUpdates().size());
-
-        // Test hex string conversion
-        byte[] bytes = HookUtils.hexStringToBytes("010203");
-        assertArrayEquals(new byte[]{0x01, 0x02, 0x03}, bytes);
-
-        String hex = HookUtils.bytesToHexString(new byte[]{0x01, 0x02, 0x03});
-        assertEquals("010203", hex);
-
-        // Test minimal representation
-        byte[] minimal = HookUtils.toMinimalRepresentation(new byte[]{0x00, 0x00, 0x01, 0x02});
-        assertArrayEquals(new byte[]{0x01, 0x02}, minimal);
-
-        // Test padding
-        byte[] padded = HookUtils.padToLength(new byte[]{0x01, 0x02}, 4);
-        assertArrayEquals(new byte[]{0x00, 0x00, 0x01, 0x02}, padded);
-    }
-
-    @Test
-    public void testStorageUtils() {
-        // Test keccak256
-        byte[] input = "hello".getBytes();
-        byte[] hash = StorageUtils.keccak256(input);
-        assertEquals(32, hash.length);
-
-        // Test mapping storage slot calculation
-        byte[] mappingSlot = new byte[32];
-        mappingSlot[0] = 0x01; // Set first byte to 1 for minimal representation
-        byte[] key = new byte[]{0x02, 0x03};
-        byte[] storageSlot = StorageUtils.calculateMappingStorageSlot(mappingSlot, key);
-        assertEquals(32, storageSlot.length);
-
-        // Test mapping storage slot with preimage
-        byte[] preimage = "test".getBytes();
-        byte[] storageSlotWithPreimage = StorageUtils.calculateMappingStorageSlotWithPreimage(mappingSlot, preimage);
-        assertEquals(32, storageSlotWithPreimage.length);
-
-        // Test encoding
-        byte[] encodedNumber = StorageUtils.encodeNumber(12345L);
-        assertTrue(encodedNumber.length > 0);
-
-        byte[] encodedString = StorageUtils.encodeString("hello");
-        assertArrayEquals("hello".getBytes(), encodedString);
-
-        // Test address encoding
-        byte[] address = new byte[20];
-        address[0] = 0x01; // Set first byte to 1 for minimal representation
-        byte[] encodedAddress = StorageUtils.encodeAddress(address);
-        assertEquals(32, encodedAddress.length);
-        assertEquals(0x01, encodedAddress[12]); // Address is padded to 32 bytes, so first byte is at index 12
-
-        // Test storage update creation
-        LambdaStorageUpdate directUpdate = StorageUtils.createDirectStorageUpdate(new byte[]{0x01}, new byte[]{0x02});
-        assertTrue(directUpdate instanceof LambdaStorageUpdate.LambdaStorageSlot);
-
-        LambdaStorageUpdate mappingUpdate = StorageUtils.createMappingStorageUpdate(mappingSlot, key, new byte[]{0x04});
-        assertTrue(mappingUpdate instanceof LambdaStorageUpdate.LambdaMappingEntries);
-
-        // Test number storage update
-        LambdaStorageUpdate numberUpdate = StorageUtils.createNumberStorageUpdate(new byte[]{0x01}, 12345L);
-        assertTrue(numberUpdate instanceof LambdaStorageUpdate.LambdaStorageSlot);
-
-        // Test string storage update
-        LambdaStorageUpdate stringUpdate = StorageUtils.createStringStorageUpdate(new byte[]{0x01}, "hello");
-        assertTrue(stringUpdate instanceof LambdaStorageUpdate.LambdaStorageSlot);
-
-        // Test address storage update (skip this test as it conflicts with minimal representation validation)
-        // LambdaStorageUpdate addressUpdate = StorageUtils.createAddressStorageUpdate(new byte[]{0x01}, address);
-        // assertTrue(addressUpdate instanceof LambdaStorageUpdate.LambdaStorageSlot);
-
-        // Test clear storage updates
-        LambdaStorageUpdate clearUpdate = StorageUtils.createClearStorageUpdate(new byte[]{0x01});
-        assertTrue(clearUpdate instanceof LambdaStorageUpdate.LambdaStorageSlot);
-        assertArrayEquals(new byte[0], ((LambdaStorageUpdate.LambdaStorageSlot) clearUpdate).getValue());
-
-        LambdaStorageUpdate clearMappingUpdate = StorageUtils.createClearMappingUpdate(mappingSlot, key);
-        assertTrue(clearMappingUpdate instanceof LambdaStorageUpdate.LambdaMappingEntries);
-    }
-
-    @Test
-    public void testStorageUtilsFromHex() {
-        // Test storage slot from hex
-        LambdaStorageUpdate storageSlot = HookUtils.createStorageSlotFromHex("0102", "0304");
-        assertTrue(storageSlot instanceof LambdaStorageUpdate.LambdaStorageSlot);
-        assertArrayEquals(new byte[]{0x01, 0x02}, ((LambdaStorageUpdate.LambdaStorageSlot) storageSlot).getKey());
-        assertArrayEquals(new byte[]{0x03, 0x04}, ((LambdaStorageUpdate.LambdaStorageSlot) storageSlot).getValue());
-
-        // Test mapping entry from hex
-        LambdaStorageUpdate mappingEntry = HookUtils.createMappingEntryFromHex("01", "0304", "0506");
-        assertTrue(mappingEntry instanceof LambdaStorageUpdate.LambdaMappingEntries);
-
-        // Test invalid hex strings
-        assertThrows(IllegalArgumentException.class, () -> {
-            HookUtils.hexStringToBytes("01G2");
-        });
-
-        assertThrows(IllegalArgumentException.class, () -> {
-            HookUtils.hexStringToBytes("0");
-        });
-    }
-
-    @Test
-    public void testHookUtilsHelpers() {
-        ContractId contractId = new ContractId(400);
-        AccountId accountId = new AccountId(500);
-
-        // Test hook ID extraction
-        HookCreationDetails hook1 = HookUtils.createSimpleAccountAllowanceHook(1L, contractId);
-        HookCreationDetails hook2 = HookUtils.createSimpleAccountAllowanceHook(2L, contractId);
-        List<Long> hookIds = HookUtils.extractHookIds(List.of(hook1, hook2));
-        assertEquals(List.of(1L, 2L), hookIds);
-
-        // Test filtering by extension point
-        List<HookCreationDetails> allHooks = List.of(hook1, hook2);
-        List<HookCreationDetails> filteredHooks = HookUtils.filterByExtensionPoint(allHooks, HookExtensionPoint.ACCOUNT_ALLOWANCE_HOOK);
-        assertEquals(2, filteredHooks.size());
-
-        // Test hook ID for entity
-        HookEntityId entityId = HookEntityId.ofAccount(accountId);
-        HookId hookId = new HookId(entityId, 1L);
-        assertTrue(HookUtils.isHookForEntity(hookId, entityId));
-
-        HookEntityId otherEntityId = HookEntityId.ofAccount(new AccountId(600));
-        assertFalse(HookUtils.isHookForEntity(hookId, otherEntityId));
-
-        // Test empty storage updates
-        List<LambdaStorageUpdate> emptyUpdates = HookUtils.emptyStorageUpdates();
-        assertTrue(emptyUpdates.isEmpty());
-
-        // Test single storage update
-        LambdaStorageUpdate singleUpdate = new LambdaStorageUpdate.LambdaStorageSlot(new byte[]{0x01}, new byte[]{0x02});
-        List<LambdaStorageUpdate> singleUpdates = HookUtils.singleStorageUpdate(singleUpdate);
-        assertEquals(1, singleUpdates.size());
-        assertEquals(singleUpdate, singleUpdates.get(0));
-    }
-
-    @Test
-    public void testStorageUtilsMultipleUpdates() {
-        // Test creating multiple mapping updates
-        byte[] mappingSlot = new byte[32];
-        mappingSlot[0] = 0x01; // Set first byte to 1 for minimal representation
-
-        LambdaStorageUpdate multipleUpdates = StorageUtils.createMappingStorageUpdates(
-                mappingSlot,
-                new byte[]{0x01}, new byte[]{0x02},
-                new byte[]{0x03}, new byte[]{0x04}
-        );
-
-        assertTrue(multipleUpdates instanceof LambdaStorageUpdate.LambdaMappingEntries);
-        LambdaStorageUpdate.LambdaMappingEntries mappingEntries = (LambdaStorageUpdate.LambdaMappingEntries) multipleUpdates;
-        assertEquals(2, mappingEntries.getEntries().size());
-
-        // Test creating storage updates from pairs
-        List<LambdaStorageUpdate> pairUpdates = HookUtils.createStorageUpdatesFromPairs(
-                new byte[]{0x01}, new byte[]{0x02},
-                new byte[]{0x03}, new byte[]{0x04}
-        );
-
-        assertEquals(2, pairUpdates.size());
-        assertTrue(pairUpdates.get(0) instanceof LambdaStorageUpdate.LambdaStorageSlot);
-        assertTrue(pairUpdates.get(1) instanceof LambdaStorageUpdate.LambdaStorageSlot);
-    }
-
-    @Test
-    public void testValidationErrors() {
-        // Test invalid address length
-        assertThrows(IllegalArgumentException.class, () -> {
-            StorageUtils.encodeAddress(new byte[19]);
-        });
-
-        // Test invalid bits for number encoding
-        assertThrows(IllegalArgumentException.class, () -> {
-            StorageUtils.encodeNumber(123L, 7);
-        });
-
-        // Test too many bits for number encoding
-        assertThrows(IllegalArgumentException.class, () -> {
-            StorageUtils.encodeNumber(123L, 264); // 33 bytes
-        });
-
-        // Test invalid key-value pairs
-        assertThrows(IllegalArgumentException.class, () -> {
-            HookUtils.createStorageUpdatesFromPairs(new byte[]{0x01});
-        });
-
-        // Test invalid mapping slot (leading zeros)
-        assertThrows(IllegalArgumentException.class, () -> {
-            StorageUtils.createMappingStorageUpdates(new byte[]{0x00, 0x01}, new byte[]{0x01}, new byte[]{0x02});
-        });
-    }
-}
